+++ openmetadata-ui/src/main/resources/ui/src/components/OntologyExplorer/hooks/useGraphData.ts
@@ -327,9 +327,20 @@ export function useGraphDataBuilder({
 
       const visibleIds = new Set([...visibleTermIds, ...visibleAssetIds]);
       nodesForGraph = inputNodes.filter((n) => visibleIds.has(n.id));
-      edgesForGraph = mergedEdgesList.filter(
-        (e) => visibleIds.has(e.from) && visibleIds.has(e.to)
-      );
+      edgesForGraph = mergedEdgesList.filter((e) => {
+        if (!visibleIds.has(e.from) || !visibleIds.has(e.to)) {
+          return false;
+        }
+        const fromIsAsset = allAssetIds.has(e.from);
+        const toIsAsset = allAssetIds.has(e.to);
+        if (fromIsAsset || toIsAsset) {
+          const termId = fromIsAsset ? e.to : e.from;
+
+          return idsToExpand.has(termId);
+        }
+
+        return true;
+      });
     } else if (explorationMode === 'hierarchy') {
       nodesForGraph = inputNodes;
       edgesForGraph = inputEdges.map((e) => ({
