+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ar-sa.json
@@ -2872,6 +2872,7 @@
     "no-recently-viewed-date": "لم تقم بعرض أي أصول بيانات مؤخرًا. استكشف للعثور على شيء مثير للاهتمام!",
     "no-reference-available": "لا توجد مراجع متاحة.",
     "no-related-terms-available": "لا توجد مصطلحات ذات صلة متاحة.",
+    "no-relations-for-selected-filter": "لم يتم العثور على علاقات لأنواع العلاقة المحددة. حاول تحديد أنواع مختلفة.",
     "no-relations-found": "لم يتم العثور على علاقات لهذا المصطلح",
     "no-retry-queue-records": "لم يتم العثور على سجلات إعادة محاولة الفهرسة المباشرة. تمت فهرسة جميع الكيانات بنجاح.",
     "no-roles-assigned": "لم يتم تعيين أدوار",
