+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/mr-in.json
@@ -2872,6 +2872,7 @@
     "no-recently-viewed-date": "तुम्ही अलीकडे कोणतेही डेटा ॲसेट पाहिलेले नाहीत. काहीतरी मनोरंजक शोधण्यासाठी एक्सप्लोर करा!",
     "no-reference-available": "कोणतेही संदर्भ उपलब्ध नाहीत.",
     "no-related-terms-available": "संबंधित संज्ञा उपलब्ध नाहीत.",
+    "no-relations-for-selected-filter": "निवडलेल्या संबंध प्रकारांसाठी कोणतेही संबंध आढळले नाहीत. वेगळे प्रकार निवडण्याचा प्रयत्न करा.",
     "no-relations-found": "या संज्ञेसाठी संबंध आढळले नाहीत",
     "no-retry-queue-records": "लाइव्ह इंडेक्सिंग पुन्हा प्रयत्न रेकॉर्ड सापडले नाहीत. सर्व एंटिटी यशस्वीरित्या इंडेक्स केल्या आहेत.",
     "no-roles-assigned": "कोणत्याही भूमिका नियुक्त केलेल्या नाहीत",
