+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ru-ru.json
@@ -2872,6 +2872,7 @@
     "no-recently-viewed-date": "Нет недавно просмотренных данных.",
     "no-reference-available": "Нет доступных ссылок.",
     "no-related-terms-available": "Нет доступных связанных терминов.",
+    "no-relations-for-selected-filter": "Отношения для выбранных типов не найдены. Попробуйте выбрать другие типы.",
     "no-relations-found": "Связи для данного термина не найдены",
     "no-retry-queue-records": "Записей повторной индексации не найдено. Все сущности успешно проиндексированы.",
     "no-roles-assigned": "Роли не назначены",
