+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/th-th.json
@@ -2872,6 +2872,7 @@
     "no-recently-viewed-date": "คุณยังไม่ได้ดูสินทรัพย์ข้อมูลใด ๆ เมื่อเร็ว ๆ นี้ สำรวจเพื่อค้นหาสิ่งที่น่าสนใจ!",
     "no-reference-available": "ไม่มีการอ้างอิงที่ใช้งานได้",
     "no-related-terms-available": "ไม่มีคำที่เกี่ยวข้องที่สามารถใช้งานได้",
+    "no-relations-for-selected-filter": "ไม่พบความสัมพันธ์สำหรับประเภทความสัมพันธ์ที่เลือก ลองเลือกประเภทอื่น",
     "no-relations-found": "ไม่พบความสัมพันธ์สำหรับคำศัพท์นี้",
     "no-retry-queue-records": "ไม่พบบันทึกการลองจัดทำดัชนีใหม่ เอนทิตีทั้งหมดได้รับการจัดทำดัชนีสำเร็จแล้ว",
     "no-roles-assigned": "ไม่มีบทบาทที่มอบหมาย",
